{"schema":"decodex/commit/1","summary":"{\"schema\":\"decodex/commit/1\",\"summary\":\"Split docs extension acceptance lifecycle tests\",\"authority\":\"manual\"}","authority":"manual"}
